add placeholder arguments in the service locator services
this argument is not needed and uses a not existing service.
at some point symfony tries to replace/merge service locators by replacing them with their hashes
and since is common service locators, it is common to have collisions.
But somehow symfony does not detect that we change the arguments if it in a compiler pass